Replacing karma+mocha with jest+enzyme.
This change is:

  * Replacing previously little used karma + mocha with jest + enzyme,
which allows us to do snapshot testing
  * Updating enzyme
  * Converting existing tests to work with jest
  * Making maven run jest for tests
  * Establishing consistent directory structure for tests (tests for
`src/foo/bar/Baz.jsx` are in `test/foo/bar/Baz.test.js`)
