+++ graylog2-server/pom.xml
@@ -781,6 +781,17 @@
                                     <arguments>run build</arguments>
                                 </configuration>
                             </execution>
+
+                            <execution>
+                                <id>npm test</id>
+                                <goals>
+                                    <goal>npm</goal>
+                                </goals>
+                                <phase>test</phase>
+                                <configuration>
+                                    <arguments>test</arguments>
+                                </configuration>
+                            </execution>
                         </executions>
                     </plugin>
                 </plugins>
