--- graylog2-web-interface/karma.ci.conf.js
@@ -1,10 +1,0 @@
-var karmaSharedConfig = require('./karma.shared');
-
-module.exports = function(config) {
-  karmaSharedConfig(config);
-  config.set({
-    autoWatch: false,
-    browsers: ['PhantomJS'],
-    singleRun: true,
-  });
-};
