--- graylog2-web-interface/karma.dev.conf.js
@@ -1,10 +1,0 @@
-var karmaSharedConfig = require('./karma.shared');
-
-module.exports = function (config) {
-  karmaSharedConfig(config);
-  config.set({
-    autoWatch: true,
-    browsers: ['Chrome'],
-    singleRun: false,
-  });
-};
