+++ graylog2-web-interface/src/components/common/Spinner.jsx
@@ -1,19 +1,13 @@
-import PropTypes from 'prop-types';
 import React from 'react';
+import PropTypes from 'prop-types';
 
-const Spinner = React.createClass({
-  propTypes: {
-    text: PropTypes.string,
-  },
-  getDefaultProps() {
-    return { text: 'Loading...' };
-  },
-  getInitialState() {
-    return {};
-  },
-  render() {
-    return (<span><i className="fa fa-spin fa-spinner" /> {this.props.text}</span>);
-  },
-});
+const Spinner = ({ text }) => <span><i className="fa fa-spin fa-spinner" /> {text}</span>;
+
+Spinner.propTypes = {
+  text: PropTypes.string,
+};
+Spinner.defaultProps = {
+  text: 'Loading...',
+};
 
 export default Spinner;
