+++ graylog2-web-interface/src/util/AppConfig.js
@@ -1,10 +1,10 @@
 const AppConfig = {
   gl2ServerUrl() {
-    return window.appConfig.gl2ServerUrl;
+    return this.appConfig().gl2ServerUrl;
   },
 
   gl2AppPathPrefix() {
-    return window.appConfig.gl2AppPathPrefix;
+    return this.appConfig().gl2AppPathPrefix;
   },
 
   gl2DevMode() {
@@ -14,7 +14,11 @@ const AppConfig = {
   },
 
   rootTimeZone() {
-    return window.appConfig.rootTimeZone;
+    return this.appConfig().rootTimeZone;
+  },
+
+  appConfig() {
+    return window.appConfig || {};
   },
 };
