--- graylog2-web-interface/test/SpecRunner.html
@@ -1,23 +1,0 @@
-<!DOCTYPE HTML>
-<html>
-<head>
-  <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-  <title>Jasmine Spec Runner v2.0.0</title>
-
-  <link rel="shortcut icon" type="image/png" href="jasmine-standalone-2.0.0/lib/jasmine-2.0.0/jasmine_favicon.png">
-  <link rel="stylesheet" type="text/css" href="jasmine-standalone-2.0.0/lib/jasmine-2.0.0/jasmine.css">
-
-  <script type="text/javascript" src="jasmine-standalone-2.0.0/lib/jasmine-2.0.0/jasmine.js"></script>
-  <script type="text/javascript" src="jasmine-standalone-2.0.0/lib/jasmine-2.0.0/jasmine-html.js"></script>
-  <script type="text/javascript" src="jasmine-standalone-2.0.0/lib/jasmine-2.0.0/boot.js"></script>
-
-  <!-- include source files here... -->
-  <!-- include spec files here... -->
-  <script type="text/javascript" src="shim/server-side-global-vars.js"></script>
-  <script type="text/javascript" src="../dist/tests.js"></script>
-
-</head>
-
-<body>
-</body>
-</html>
