--- graylog2-web-interface/test/jasmine-standalone-2.0.0/src/Song.js
@@ -1,7 +1,0 @@
-function Song() {
-}
-
-Song.prototype.persistFavoriteStatus = function(value) {
-  // something complicated
-  throw new Error("not yet implemented");
-};
