--- graylog2-web-interface/test/shim/server-side-global-vars.js
@@ -1,1 +1,0 @@
-var gl2AppPathPrefix = "";
